upsample mode to enum file
