+++ src/super_gradients/common/__init__.py
@@ -4,8 +4,8 @@ from super_gradients.common.aws_connection import AWSConnector
 from super_gradients.common.data_connection import S3Connector
 from super_gradients.common.data_interface import DatasetDataInterface, ADNNModelRepositoryDataInterfaces
 from super_gradients.common.environment.env_helpers import init_trainer, is_distributed
-from super_gradients.common.data_types import StrictLoad, DeepLearningTask, EvaluationType, MultiGPUMode
+from super_gradients.common.data_types import StrictLoad, DeepLearningTask, EvaluationType, MultiGPUMode, UpsampleMode
 
 __all__ = ['explicit_params_validation', 'singleton', 'AWSConnector', 'DatasetDataInterface',
            'ADNNModelRepositoryDataInterfaces', 'S3Connector', 'init_trainer', 'is_distributed',
-           'StrictLoad', 'DeepLearningTask', 'EvaluationType', 'MultiGPUMode']
+           'StrictLoad', 'DeepLearningTask', 'EvaluationType', 'MultiGPUMode', 'UpsampleMode']
