+++ src/super_gradients/common/data_types/__init__.py
@@ -1,3 +1,4 @@
-from super_gradients.common.data_types.enum import StrictLoad, DeepLearningTask, MultiGPUMode, EvaluationType
+from super_gradients.common.data_types.enum import StrictLoad, DeepLearningTask, MultiGPUMode, EvaluationType,\
+    UpsampleMode
 
-__all__ = ['StrictLoad', 'DeepLearningTask', 'EvaluationType', 'MultiGPUMode']
+__all__ = ['StrictLoad', 'DeepLearningTask', 'EvaluationType', 'MultiGPUMode', 'UpsampleMode']
