+++ src/super_gradients/common/data_types/enum/__init__.py
@@ -2,6 +2,7 @@ from super_gradients.common.data_types.enum.strict_load import StrictLoad
 from super_gradients.common.data_types.enum.deep_learning_task import DeepLearningTask
 from super_gradients.common.data_types.enum.evaluation_type import EvaluationType
 from super_gradients.common.data_types.enum.multi_gpu_mode import MultiGPUMode
+from super_gradients.common.data_types.enum.upsample_mode import UpsampleMode
 
 
-__all__ = ['StrictLoad', 'DeepLearningTask', 'EvaluationType', 'MultiGPUMode']
+__all__ = ['StrictLoad', 'DeepLearningTask', 'EvaluationType', 'MultiGPUMode', 'UpsampleMode']
