+++ src/super_gradients/common/data_types/enum/upsample_mode.py
@@ -1,0 +1,8 @@
+from enum import Enum
+
+
+class UpsampleMode(Enum):
+    NEAREST = "nearest"
+    BILINEAR = "bilinear"
+    BICUBIC = "bicubic"
+    SNPE_BILINEAR = "snpe_bilinear"
