+++ src/super_gradients/training/models/segmentation_models/ppliteseg.py
@@ -2,8 +2,11 @@ import torch
 import torch.nn as nn
 import torch.nn.functional as F
 from typing import Union, List, Tuple
-from super_gradients.training.utils.module_utils import ConvBNReLU, UpsampleMode, make_upsample_module
-from super_gradients.training.models.segmentation_models.stdc import SegmentationHead, AbstractSTDCBackbone, STDC1Backbone, STDC2Backbone
+
+from super_gradients.training.utils.module_utils import ConvBNReLU, make_upsample_module
+from super_gradients.common import UpsampleMode
+from super_gradients.training.models.segmentation_models.stdc import SegmentationHead, AbstractSTDCBackbone,\
+    STDC1Backbone, STDC2Backbone
 from super_gradients.training.models.segmentation_models.segmentation_module import SegmentationModule
 from super_gradients.training.utils import HpmStruct, get_param
